feat: rename History to Activities and fix filter icon
- Rename bottom nav tab from "History" to "Activities"
- Replace gear icon (⚙) with SVG filter/funnel icon
- Add aria-expanded to filter toggle button
- Increase .btn-icon touch target from 36px to 44px (WCAG 2.5.8)
- Rename components/history/ → components/activities/
- Update route name, CSS classes, imports, and tests

Refs #11
