+++ CLAUDE.md
@@ -53,7 +53,7 @@ Comma-separated, multiple per exercise. Common tags: Push, Pull, Legs, Chest, Ba
 ## UX Design Decisions
 These decisions were made with the user and must be respected by all agents:
 
-- **Landing screen**: Workout History (chronological list with floating "Start Workout" button)
+- **Landing screen**: Activities (chronological workout list with floating "Start Workout" button)
 - **Set logging**: Logbook-style — record sets when convenient, not real-time per-set
 - **Rest timers**: None in-app — rest times are planning data only (user uses watch)
 - **Supersets**: Flat list with section labels/colors (not grouped flow)
@@ -61,9 +61,9 @@ These decisions were made with the user and must be respected by all agents:
 - **Exercise library**: Inline creation (on-the-fly during workout/template building)
 - **Set entry**: Quick-fill weight for all sets, override individual sets if different
 - **Templates**: Create via dedicated editor AND save-from-workout
-- **Navigation**: Bottom tab bar (History | Templates | Settings)
-- **History cards**: Compact (Date, Name, Type badge) — tap to view details
-- **History detail**: Summary per exercise, tap to expand set-by-set
+- **Navigation**: Bottom tab bar (Activities | Templates | Exercises | Settings)
+- **Activity cards**: Compact (Date, Name, Type badge) — tap to view details
+- **Activity detail**: Summary per exercise, tap to expand set-by-set
 - **Copy workout**: Pre-fill from previous + show "last time" reference while logging
 - **Device**: Mobile-first (375px primary breakpoint)
