+++ frontend/src/components/activities/exercise-detail.test.ts
