+++ frontend/src/components/activities/exercise-detail.tsx
